requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
algoliasearch=2.5.0=pypi_0
astroid=2.6.6=py39hecd8cb5_0
blas=1.0=mkl
bottleneck=1.3.2=py39he3068b8_1
ca-certificates=2021.7.5=hecd8cb5_1
certifi=2021.5.30=py39hecd8cb5_0
charset-normalizer=2.0.4=pypi_0
dnspython=1.16.0=py39hecd8cb5_0
docopt=0.6.2=pypi_0
idna=3.2=pypi_0
intel-openmp=2021.3.0=hecd8cb5_3375
isort=5.9.3=pyhd3eb1b0_0
lazy-object-proxy=1.6.0=py39h9ed2024_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
mccabe=0.6.1=py39hecd8cb5_1
mkl=2021.3.0=hecd8cb5_517
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.0=py39h4a7008c_2
mkl_random=1.2.2=py39hb2f4e1b_0
ncurses=6.2=h0a44026_1
numexpr=2.7.3=py39h5873af2_1
numpy=1.20.3=py39h4b4dc7a_0
numpy-base=1.20.3=py39he0bd621_0
openssl=1.1.1l=h9ed2024_0
pandas=1.3.2=py39h5008ddb_0
pip=21.2.4=py37hecd8cb5_0
pipreqs=0.4.10=pypi_0
pylint=2.9.6=py39hecd8cb5_1
pymongo=3.12.0=py39h23ab428_0
python=3.9.6=h88f2d9e_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.19.0=pypi_0
pytz=2021.1=pyhd3eb1b0_0
readline=8.1=h9ed2024_0
requests=2.26.0=pypi_0
setuptools=52.0.0=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hce871da_0
tabulate=0.8.9=py39hecd8cb5_0
tk=8.6.10=hb0a8c7a_0
toml=0.10.2=pyhd3eb1b0_0
tzdata=2021a=h5d7bf9c_0
urllib3=1.26.6=pypi_0
wheel=0.37.0=pyhd3eb1b0_0
wrapt=1.12.1=py39h9ed2024_1
xz=5.2.5=h1de35cc_0
yarg=0.1.9=pypi_0
zlib=1.2.11=h1de35cc_3
